chore #8: style 파일 분리
