+++ trecord/src/components/Comment/CommentReplyBtn/style.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+export const Layout = styled.button`
+  ${({ theme }) => theme.font.fontSize.Caption_S}
+  ${({ theme }) => theme.font.fontType.R};
+  background: none;
+  border: none;
+  display: flex;
+  padding: 0;
+`;
