+++ trecord/src/components/common/Editor/index.tsx
@@ -1,54 +1,8 @@
 import { ReactElement, useMemo, useRef } from 'react';
-import styled from 'styled-components';
 import 'react-quill/dist/quill.snow.css';
 import { uploadS3 } from '@/utils/image';
 import ReactQuill from 'react-quill';
-
-const StyledContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-  min-height: 85vh;
-
-  .ql-toolbar {
-    border: 1px solid ${({ theme }) => theme.colors.colorStyles.gray300};
-    border-right: none;
-    border-left: none;
-    display: inline-flex;
-    justify-content: end;
-
-    width: 100%;
-    @media (min-width: 431px) {
-      width: 390px;
-    }
-    background: ${({ theme }) => theme.colors.colorStyles.gray100};
-    z-index: 1;
-    height: 42px;
-    position: fixed;
-  }
-
-  .ql-container {
-    border: none;
-    padding-top: 42px;
-    height: 75vh;
-    overflow: auto;
-
-    scrollbar-width: none;
-    ::-webkit-scrollbar {
-      display: none;
-    }
-  }
-
-  .ql-editor::before {
-    color: ${({ theme }) => theme.colors.colorStyles.gray600};
-    font-family: Pretendard;
-    font-size: 16px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 24px; /* 150% */
-  }
-`;
-
+import * as S from './style';
 const Editor = ({
   content,
   contentSetter: setContent,
@@ -97,7 +51,7 @@ const Editor = ({
   );
 
   return (
-    <StyledContainer>
+    <S.Layout>
       <ReactQuill
         ref={quillRef}
         theme="snow"
@@ -106,7 +60,7 @@ const Editor = ({
         value={content}
         onChange={setContent}
       />
-    </StyledContainer>
+    </S.Layout>
   );
 };
