+++ trecord/src/components/common/Toast/index.tsx
@@ -1,6 +1,5 @@
 import React, { useState, ReactNode } from 'react';
-import styled from 'styled-components';
-
+import * as S from './style';
 interface ToastContextProps {
   showToast: (message: string, duration?: number) => void;
 }
@@ -8,44 +7,6 @@ interface ToastContextProps {
 export const ToastContext = React.createContext<ToastContextProps | undefined>(
   undefined,
 );
-
-const StyledToastContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: end;
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  pointer-events: none; /* So that it doesn't block other UI elements */
-  z-index: 1500;
-  padding: 10%;
-  @media (min-width: 431px) {
-    padding: 19vh;
-  }
-
-  .toast {
-    padding: 12px 16px;
-    border-radius: 4px;
-    background-color: rgba(0, 0, 0, 0.87);
-    color: ${({ theme }) => theme.colors.colorStyles.gray100};
-    font-size: 16px;
-    max-width: 90%;
-    word-wrap: break-word;
-    box-shadow: 0 3px 5px rgba(0, 0, 0, 0.2);
-    transition:
-      opacity 0.3s,
-      transform 0.3s;
-    transform: translateY(0);
-    opacity: 1;
-  }
-
-  .toast.exiting {
-    transform: translateY(100px);
-    opacity: 0;
-  }
-`;
 
 interface ToastProps {
   message: string;
@@ -68,11 +29,11 @@ export const ToastProvider: React.FC<{ children: ReactNode }> = ({
     <ToastContext.Provider value={{ showToast }}>
       {children}
       {toast && (
-        <StyledToastContainer>
+        <S.Layout>
           <div className={`toast ${toast ? '' : 'exiting'}`}>
             {toast.message}
           </div>
-        </StyledToastContainer>
+        </S.Layout>
       )}
     </ToastContext.Provider>
   );
