+++ trecord/src/components/common/skeleton/index.tsx
@@ -1,29 +1,5 @@
 import { ReactElement } from 'react';
-import styled from 'styled-components';
-
-interface Props {
-  width?: string;
-  height?: string;
-}
-
-const StyledSkeleton = styled.div<Props>`
-  background: linear-gradient(90deg, #f0f0f0 25%, #e0e0e0 50%, #f0f0f0 75%);
-  background-size: 200% 100%;
-  animation: shimmer 1.5s infinite;
-  width: ${({ width = '90%' }) => width};
-  height: ${({ height = '160px' }) => height};
-  min-height: ${({ height = '160px' }) => height};
-  border-radius: 10px;
-
-  @keyframes shimmer {
-    0% {
-      background-position: -200% 0;
-    }
-    100% {
-      background-position: 200% 0;
-    }
-  }
-`;
+import * as S from './style';
 
 interface SkeletonProps {
   num?: number;
@@ -42,7 +18,7 @@ const Skeleton = ({
         .fill(0)
         .map((x, i) => x + i)
         .map((y) => (
-          <StyledSkeleton key={y} width={width} height={height} />
+          <S.Layout key={y} width={width} height={height} />
         ))}
     </>
   );
