+++ trecord/src/pages/ModifyRecord/index.tsx
@@ -3,7 +3,7 @@ import { ReactElement, useEffect, useState } from 'react';
 import { useLocation, useNavigate, useParams } from 'react-router-dom';
 import { observer } from 'mobx-react-lite';
 import { useStore } from '@/stores';
-import styled from 'styled-components';
+import * as S from './style';
 import { NavBarNew } from '@components/common/navBar/NavBarNew';
 import { useGetRecord } from '@/apis';
 import ImgInput from '@components/common/ImgInput';
@@ -13,39 +13,6 @@ import { NewFeel } from '@components/NewRecord/NewFirstRecord/NewFeel';
 import { NewMove } from '@components/NewRecord/NewFirstRecord/NewMove';
 import { SquareButton } from '@components/common/button/SquareButton';
 import { AutoCompletePlace } from '@components/common/AutoCompletePlace';
-
-const StyledFrame = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 19px;
-  overflow: auto;
-  padding: 0px 20px;
-  box-sizing: border-box;
-
-  padding-top: 100px;
-  gap: 19px;
-  @media (min-width: 431px) {
-    height: calc(844px - 120px);
-  }
-  width: 100%;
-  scrollbar-width: none;
-  align-items: center;
-  padding-bottom: 20px;
-  height: calc(100% - 10px);
-  ::-webkit-scrollbar {
-    display: none;
-  }
-  .new_feel {
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    width: 356px;
-    gap: 15px;
-  }
-  .new_btn {
-    padding-top: 30px;
-  }
-`;
 
 const ModifyRecord = observer((): ReactElement => {
   const { id: recordId = '' } = useParams();
@@ -125,7 +92,7 @@ const ModifyRecord = observer((): ReactElement => {
           navigate(-1);
         }}
       />
-      <StyledFrame>
+      <S.Layout>
         <ImgInput imgFile={thumbNail} imgFileSetter={setThumbNail} />
         <TextInput
           inputValue={title}
@@ -163,7 +130,7 @@ const ModifyRecord = observer((): ReactElement => {
             onClick={handleClickNext}
           />
         </div>
-      </StyledFrame>
+      </S.Layout>
     </>
   );
 });
